Commit - ForgotPasswordPage Exception Handeling and Validator
Commit - ForgotPasswordPage Exception Handeling and Validator

:white_check_mark:

* the MyTextField widget does not accept a validator parameter in its current implementation. To fix this issue, you need to modify the MyTextField widget to accept a validator parameter and use the TextFormField instead of TextField.

* This code adds a validator parameter to the MyTextField widget and replaces the TextField widget with the TextFormField widget, which allows you to use the validator parameter for input validation. Now your ForgotPasswordPage should work without any issues.

---

In the updated MyTextField widget I provided earlier, the validator parameter is optional because it has a default value of null. This means you can use the MyTextField widget without providing a validator, and it will not perform any validation in that case.

* int this example :

https://pastebin.com/raw/Eq4MXUud

The first MyTextField instance has a validator function to validate email input, while the second one does not have any validation because the validator parameter is not provided. The second MyTextField instance will work without any issues since the validator parameter is optional.